Tip a result with a normally distributed confounder.

choose one of the following, and the other will be estimated:

smd

outcome_association

Usage

tip(
  effect,
  smd = NULL,
  outcome_association = NULL,
  verbose = TRUE,
  correction_factor = "none"
)

tip_with_continuous(
  effect,
  smd = NULL,
  outcome_association = NULL,
  verbose = TRUE,
  correction_factor = "none"
)

tip_c(
  effect,
  smd = NULL,
  outcome_association = NULL,
  verbose = TRUE,
  correction_factor = "none"
)

Arguments

effect

Numeric positive value. Observed exposure - outcome effect (assumed to be the exponentiated coefficient, so a relative risk, odds ratio, or hazard ratio). This can be the point estimate, lower confidence bound, or upper confidence bound.

smd

Numeric. Estimated difference in scaled means between the unmeasured confounder in the exposed population and unexposed population

outcome_association

Numeric positive value. Estimated association between the unmeasured confounder and the outcome

verbose

Logical. Indicates whether to print informative message. Default: TRUE

correction_factor

Character string. Options are "none", "hr", "or". For common outcomes (>15%), the odds ratio or hazard ratio is not a good estimate for the relative risk. In these cases, we can apply a correction factor. If you are supplying a hazard ratio for a common outcome, set this to "hr"; if you are supplying an odds ratio for a common outcome, set this to "or"; if you are supplying a risk ratio or your outcome is rare, set this to "none" (default).

Value

Data frame.

Examples

## to estimate the association between an unmeasured confounder and outcome
## needed to tip analysis
tip(1.2, smd = -2)
#> The observed effect (1.2) WOULD be tipped by 1 unmeasured confounder
#> with the following specifications:
#>   * estimated difference in scaled means between the unmeasured confounder
#>     in the exposed population and unexposed population: -2
#>   * estimated association between the unmeasured confounder and the outcome: 0.91
#> 
#> # A tibble: 1 × 5
#>   effect_adjusted effect_observed   smd outcome_association n_unmeasured_confou…
#>             <dbl>           <dbl> <dbl>               <dbl>                <dbl>
#> 1               1             1.2    -2               0.913                    1

## to estimate the number of unmeasured confounders specified needed to tip
## the analysis
tip(1.2, smd = -2, outcome_association = .99)
#> The observed effect (1.2) WOULD be tipped by 9 unmeasured confounders
#> with the following specifications:
#>   * estimated difference in scaled means between the unmeasured confounder
#>     in the exposed population and unexposed population: -2
#>   * estimated association between the unmeasured confounder and the outcome: 0.99
#> 
#> # A tibble: 1 × 5
#>   effect_adjusted effect_observed   smd outcome_association n_unmeasured_confou…
#>             <dbl>           <dbl> <dbl>               <dbl>                <dbl>
#> 1               1             1.2    -2                0.99                 9.07

## Example with broom
if (requireNamespace("broom", quietly = TRUE) &&
    requireNamespace("dplyr", quietly = TRUE)) {
  glm(am ~ mpg, data = mtcars, family = "binomial") %>%
   broom::tidy(conf.int = TRUE, exponentiate = TRUE) %>%
   dplyr::filter(term == "mpg") %>%
   dplyr::pull(conf.low) %>%
   tip(outcome_association = 2.5)
}
#> The observed effect (1.13) WOULD be tipped by 1 unmeasured confounder
#> with the following specifications:
#>   * estimated difference in scaled means between the unmeasured confounder
#>     in the exposed population and unexposed population: 0.13
#>   * estimated association between the unmeasured confounder and the outcome: 2.5
#> 
#> # A tibble: 1 × 5
#>   effect_adjusted effect_observed   smd outcome_association n_unmeasured_confou…
#>             <dbl>           <dbl> <dbl>               <dbl>                <dbl>
#> 1               1            1.13 0.133                 2.5                    1