Tip a result with a binary confounder.

Choose two of the following three to specify, and the third will be estimated:

exposed_p

unexposed_p

outcome_association

Alternatively, specify all three and the function will return the number of unmeasured confounders specified needed to tip the analysis.

Usage

tip_with_binary(
  effect,
  exposed_p = NULL,
  unexposed_p = NULL,
  outcome_association = NULL,
  verbose = TRUE,
  correction_factor = "none"
)

tip_b(
  effect,
  exposed_p = NULL,
  unexposed_p = NULL,
  outcome_association = NULL,
  verbose = TRUE,
  correction_factor = "none"
)

Arguments

effect

Numeric positive value. Observed exposure - outcome effect (assumed to be the exponentiated coefficient, so a relative risk, odds ratio, or hazard ratio). This can be the point estimate, lower confidence bound, or upper confidence bound.

exposed_p

Numeric between 0 and 1. Estimated prevalence of the unmeasured confounder in the exposed population

unexposed_p

Numeric between 0 and 1. Estimated prevalence of the unmeasured confounder in the unexposed population

outcome_association

Numeric positive value. Estimated association between the unmeasured confounder and the outcome

verbose

Logical. Indicates whether to print informative message. Default: TRUE

correction_factor

Character string. Options are "none", "hr", "or". For common outcomes (>15%), the odds ratio or hazard ratio is not a good estimate for the relative risk. In these cases, we can apply a correction factor. If you are supplying a hazard ratio for a common outcome, set this to "hr"; if you are supplying an odds ratio for a common outcome, set this to "or"; if you are supplying a risk ratio or your outcome is rare, set this to "none" (default).

Details

tip_b() is an alias for tip_with_binary().

Examples

## to estimate the association between an unmeasured confounder and outcome
## needed to tip analysis
tip_with_binary(1.2, exposed_p = 0.5, unexposed_p = 0)
#> The observed effect (1.2) WOULD be tipped by 1 unmeasured confounder
#> with the following specifications:
#>   * estimated prevalence of the unmeasured confounder in the exposed population: 0.5
#>   * estimated prevalence of the unmeasured confounder in the unexposed population: 0
#>   * estimated association between the unmeasured confounder and the outcome: 1.4
#> 
#> # A tibble: 1 × 6
#>   effect_adjusted effect_observed exposed_p unexposed_p outcome_associ…¹ n_unm…²
#>             <dbl>           <dbl>     <dbl>       <dbl>            <dbl>   <dbl>
#> 1               1             1.2       0.5           0              1.4       1
#> # … with abbreviated variable names ¹​outcome_association,
#> #   ²​n_unmeasured_confounders

## to estimate the number of unmeasured confounders specified needed to tip
## the analysis
tip_with_binary(1.2,
  exposed_p = 0.5,
  unexposed_p = 0,
  outcome_association = 1.1)
#> The observed effect (1.2) WOULD be tipped by 4 unmeasured confounders
#> with the following specifications:
#>   * estimated prevalence of the unmeasured confounder in the exposed population: 0.5
#>   * estimated prevalence of the unmeasured confounder in the unexposed population: 0
#>   * estimated association between the unmeasured confounder and the outcome: 1.1
#> 
#> # A tibble: 1 × 6
#>   effect_adjusted effect_observed exposed_p unexposed_p outcome_associ…¹ n_unm…²
#>             <dbl>           <dbl>     <dbl>       <dbl>            <dbl>   <dbl>
#> 1               1             1.2       0.5           0              1.1    3.74
#> # … with abbreviated variable names ¹​outcome_association,
#> #   ²​n_unmeasured_confounders

## Example with broom
if (requireNamespace("broom", quietly = TRUE) &&
    requireNamespace("dplyr", quietly = TRUE)) {
  glm(am ~ mpg, data = mtcars, family = "binomial") %>%
   broom::tidy(conf.int = TRUE, exponentiate = TRUE) %>%
   dplyr::filter(term == "mpg") %>%
   dplyr::pull(conf.low) %>%
   tip_with_binary(exposed_p = 1, outcome_association = 1.15)
}
#> The observed effect (1.13) WOULD be tipped by 1 unmeasured confounder
#> with the following specifications:
#>   * estimated prevalence of the unmeasured confounder in the exposed population: 1
#>   * estimated prevalence of the unmeasured confounder in the unexposed population: 0.12
#>   * estimated association between the unmeasured confounder and the outcome: 1.15
#> 
#> # A tibble: 1 × 6
#>   effect_adjusted effect_observed exposed_p unexposed_p outcome_associ…¹ n_unm…²
#>             <dbl>           <dbl>     <dbl>       <dbl>            <dbl>   <dbl>
#> 1               1            1.13         1       0.119             1.15       1
#> # … with abbreviated variable names ¹​outcome_association,
#> #   ²​n_unmeasured_confounders